+++ src/Repositories/ICountRepositoryBase.cs
@@ -1,0 +1,12 @@
+﻿using PowerUtils.BuildingBlocks.Domain;
+
+namespace PowerUtils.BuildingBlocks.Data.Repositories
+{
+    public interface ICountRepositoryBase<TEntity, TId> :
+        IEntityRepositoryBase<TEntity, TId>
+
+        where TEntity : IAggregateRoot<TId>
+    {
+        long Count();
+    }
+}