+++ src/Repositories/IGetRepositoryBase.cs
@@ -1,0 +1,12 @@
+﻿using PowerUtils.BuildingBlocks.Domain;
+
+namespace PowerUtils.BuildingBlocks.Data.Repositories
+{
+    public interface IGetRepositoryBase<TEntity, TId> :
+        IEntityRepositoryBase<TEntity, TId>
+
+        where TEntity : IAggregateRoot<TId>
+    {
+        TEntity GetById(TId id);
+    }
+}