+++ src/Repositories/IListRepositoryBase.cs
@@ -1,0 +1,13 @@
+﻿using System.Collections.Generic;
+using PowerUtils.BuildingBlocks.Domain;
+
+namespace PowerUtils.BuildingBlocks.Data.Repositories
+{
+    public interface IListRepositoryBase<TEntity, TId> :
+        IEntityRepositoryBase<TEntity, TId>
+
+        where TEntity : IAggregateRoot<TId>
+    {
+        IEnumerable<TEntity> List();
+    }
+}