+++ src/Repositories/IUpdateRepositoryBase.cs
@@ -1,0 +1,12 @@
+﻿using PowerUtils.BuildingBlocks.Domain;
+
+namespace PowerUtils.BuildingBlocks.Data.Repositories
+{
+    public interface IUpdateRepositoryBase<TEntity, TId> :
+        IEntityRepositoryBase<TEntity, TId>
+
+        where TEntity : IAggregateRoot<TId>
+    {
+        void Update(TEntity entity);
+    }
+}