+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# [2.1.0](https://github.com/TechNobre/PowerUtils.BuildingBlocks.Domain.Data/compare/v2.0.0...v2.1.0) (2022-07-15)
+
+
+### Features
+
+* Add support to debug in runtime `Microsoft.SourceLink.GitHub` ([6321dea](https://github.com/TechNobre/PowerUtils.BuildingBlocks.Domain.Data/commit/6321dea4d013cca05728e800114efdf834c3797b))
+
 # [2.0.0](https://github.com/TechNobre/PowerUtils.BuildingBlocks.Domain.Data/compare/v1.0.0...v2.0.0) (2022-05-01)